+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # Changelog
 
+<h4>7..1</h4>
+(Also applies to 7.1.0)
+
+- Fix several critical bugs in the 7.x release
+- Try to fix offline install
+- Misc refactoring
+
 <h4>7.0.1</h4>
 
 - Rebrand the module to reflect new abilities